переделал на один класс
