+++ src/ru/avalon/vergentev/j120/labwork4a/Main.java
@@ -2,6 +2,7 @@ package ru.avalon.vergentev.j120.labwork4a;
 
 public class Main {
     public static void main(String[] args) {
-        new BooksAccounting().setVisible(true);
+//        new BooksAccounting().setVisible(true);
+        new BookAccounting().setVisible(true);
     }
 }
